requirements.txt
opencv-python-headless
tensorflow==2.4.1
Keras==2.4.3
numpy==1.19.5
aiortc==1.2.0
av==8.0.3
streamlit==0.80.0
Pillow==8.2.0
streamlit_webrtc==0.11.0
